Extract MqttCommandHandler
